Bug 1470362 - remove secret value from the environment
When tests are run on a desktop, typically things like
TASKCLUSTER_CLIENT_ID are defined as environment variables.  When they
are run in automation, these are fetched from the secrets service, and
thus not available in environment variables.

It's not uncommon to see tests pass on a desktop and not in automation
because of this difference.  This is a real bug in the software:
somewhere, something is looking at the environment variables instead of
the configuration.

Deleting the environment variables will surface this error sooner,
during development on the desktop.